Refactor added simplecoV and shotgun gems
