--- app/controllers/application_controller.rb
@@ -1,4 +1,0 @@
-require 'sinatra'
-
-class ApplicationController < Sinatra::Base
-end
