+++ config.ru
@@ -1,5 +1,9 @@
-require 'sinatra'
+require 'sinatra/base'
 require 'bundler'
-require './zipcode_base_api'
+# require_relative 'zipcode_base_api'
+require './app/controllers/distance_controller'
+require './app/controllers/radius_controller'
+require './app/controllers/zipcode_controller'
 Bundler.require
 run Sinatra::Application
+
